python-functions/requirements.txt
scikit-learn
pandas
numpy
supabase
joblib
python-dotenv
alpaca-trade-api
fastapi
uvicorn
pydantic

